refactor: using uint instead of int64

+++ recommend.go
@@ -3,30 +3,30 @@ package lightnovel
 import "net/http"
 
 type RecommendItem struct {
-	Gid        int64       `json:"gid"`
-	Type       int64       `json:"type"`
+	Gid        uint        `json:"gid"`
+	Type       uint        `json:"type"`
 	Title      string      `json:"title"`
-	Rows       int64       `json:"rows"`
-	Columns    int64       `json:"columns"`
+	Rows       uint        `json:"rows"`
+	Columns    uint        `json:"columns"`
 	More       string      `json:"more"`
-	MoreType   int64       `json:"more_type"`
+	MoreType   uint        `json:"more_type"`
 	MoreParams any         `json:"more_params"` // can be null string "" or uint
 	Items      []BookItems `json:"items"`
 }
 
 type BookItems struct {
-	Id              int64  `json:"id"`
-	Type            int64  `json:"type"`
+	Id              uint   `json:"id"`
+	Type            uint   `json:"type"`
 	Title           string `json:"title"`
-	ActionType      int64  `json:"action_type"`
-	ActionParams    int64  `json:"action_params"`
+	ActionType      uint   `json:"action_type"`
+	ActionParams    uint   `json:"action_params"`
 	PictureUrl      string `json:"pic_url"`
-	Gid             int64  `json:"gid"`
+	Gid             uint   `json:"gid"`
 	GroupName       string `json:"group_name"`
-	ParentGid       int64  `json:"parent_gid"`
+	ParentGid       uint   `json:"parent_gid"`
 	ParentGroupName string `json:"parent_group_name"`
-	Comments        int64  `json:"comments"`
-	Hits            int64  `json:"hits"`
+	Comments        uint   `json:"comments"`
+	Hits            uint   `json:"hits"`
 }
 
 type RecommendRequest struct {
